# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  handler.RoleResponse:
    description: Role data
    properties:
      abilities:
        example:
        - eliminate
        items:
          type: string
        type: array
      description:
        example: The mafia member eliminates villagers at night.
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Mafia
        type: string
      slug:
        example: mafia
        type: string
      team:
        enum:
        - mafia
        - village
        - independent
        example: mafia
        type: string
      video:
        example: https://example.com/video.mp4
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mafia Night Support
    url: https://github.com/mafia-night
  description: REST API for the Mafia Night game platform.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Mafia Night API
  version: "1.0"
paths:
...
  /admin/roles:
    post:
      consumes:
      - application/json
      description: Creates a new game role. Requires admin authentication.
      parameters:
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          properties:
            abilities:
              items:
                type: string
              type: array
            description:
              type: string
            name:
              type: string
            slug:
              type: string
            team:
              type: string
            video:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - admin-roles
  /admin/roles/{id}:
    delete:
      description: Deletes a role. Requires admin authentication.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - admin-roles
    patch:
      consumes:
      - application/json
      description: Updates an existing role. Requires admin authentication.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Partial role data
        in: body
        name: body
        schema:
          properties:
            abilities:
              items:
                type: string
              type: array
            description:
              type: string
            name:
              type: string
            slug:
              type: string
            team:
              type: string
            video:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - admin-roles
...
  /roles:
    get:
      description: Retrieves all available roles in the game.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RoleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      summary: List all roles
      tags:
      - roles
  /roles/{slug}:
    get:
      description: Retrieves a single role by its slug identifier.
      parameters:
      - description: Role slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      summary: Get a role by slug
      tags:
      - roles
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"